GetStudentRecentLessonsProgressQueryHandler
Has been improved

+++ TransactionDomain/Features/Tracker/DTO/Query/StudentRecentLessonProgressResponse.cs
@@ -4,6 +4,6 @@ public class StudentRecentLessonProgressResponse
 {
     public string LessonName { get; set; }
     public int LessonPoints { get; set; }
-    public float StudentPoints { get; set; }
+    public int StudentPoints { get; set; }
     public double Progress { get => Math.Round(StudentPoints / (double)LessonPoints, 2); }
 }
